+++ src/misbot/server/schemas.py
@@ -1,7 +1,8 @@
+from datetime import datetime
 from typing import Optional
 from uuid import UUID
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 
 class Player(BaseModel):
@@ -25,3 +26,9 @@ class PlayerPlayTimeResponse(BaseModel):
     hours: int
     minutes: int
     seconds: int
+
+
+class GetPlayerResponse(BaseModel):
+    player_id: UUID = Field(alias="id")
+    nickname: str | None = None
+    last_seen: datetime = Field(alias="seen")